Fixed Small Bug on last episode of a series
