+++ YourFlix Setup/YourFlix-MKIII-Server.code-workspace
@@ -1,0 +1,7 @@
+{
+	"folders": [
+		{
+			"path": "."
+		}
+	]
+}
